Do not export profile/bashrc guard vars
Some facts:

- My .profile loads .bashrc at the end, if executed by bash,
  because `bash -l` will read .profile
- My .bashrc loads .profile at the beginning,
  because `bash -i` will read .bashrc
- To avoid circular dot-sourcing, I use guard variables:
  $MRL_PROFILE_GUARD / $MRL_BASHRC_GUARD
- Defining a variable makes it available in current process
- Exporting a variable makes it available to subprocesses

Clearly, defining but not exporting the guard variables is correct.

I discovered this problem when starting to use tmux.
It appears to run shells as interactive nonlogin shells,
which read .bashrc.

However, because $MRL_PROFILE_GUARD was exported,
it will not dot-source .profile first, which leads to problems.

This commit fixes that issue